chanstate: move kv open channel put-get test
Move OpenChannel put, fetch, revocation update, and close coverage from channeldb into chanstate because these paths belong to the KV channel-state store.

Reuse the existing chanstate test fixtures and add small HTLC option helpers rather than keeping the ChannelStateDB fixture around.
